svtr_tiny GradSampler2D算子优化

+++ mindocr/models/transforms/tps_spatial_transformer.py
@@ -11,7 +11,7 @@ from mindspore import Tensor
 
 def grid_sample(input: Tensor, grid: Tensor, canvas: Optional[Tensor] = None) -> Tensor:
     out_type = input.dtype
-    output = ops.grid_sample(input.astype(ms.float64), grid.astype(ms.float64)).astype(out_type)
+    output = ops.grid_sample(input.astype(ms.float32), grid.astype(ms.float32)).astype(out_type)
     if canvas is None:
         return output
     else:
